Boxarts for ScummVM
I clicked on "upload" and it took me straight to the upload screen without asking me which folder I wanted to upload the files to. Hope this is ok. They need to go into the ScummVM folder's subfolder, "Named_Boxarts".

I noticed that in the ScummVM database, there aren't any named entries for special versions of games, e.g. the CD-ROM or FM-Towns versions of Loom. Perhaps these need to be changed?